# Conflicts:
#	android/Staccato_AN/app/src/main/java/com/on/staccato/presentation/category/viewmodel/CategoryViewModel.kt
#	android/Staccato_AN/app/src/main/res/values/strings.xml

+++ android/Staccato_AN/app/src/main/java/com/on/staccato/presentation/category/viewmodel/CategoryViewModel.kt
@@ -51,8 +51,8 @@ class CategoryViewModel
         private val _exceptionState = MutableSingleLiveData<ExceptionState2>()
         val exceptionState: SingleLiveData<ExceptionState2> get() = _exceptionState
 
-        private val _isDeleteSuccess = MutableSingleLiveData(false)
-        val isDeleteSuccess: SingleLiveData<Boolean> get() = _isDeleteSuccess
+        private val _isDeleted = MutableSingleLiveData<Boolean>()
+        val isDeleted: SingleLiveData<Boolean> get() = _isDeleted
 
         private var _isInviteMode = MutableStateFlow(false)
         val isInviteMode: StateFlow<Boolean> get() = _isInviteMode
@@ -127,7 +127,12 @@ class CategoryViewModel
 
         fun deleteCategory() {
             viewModelScope.launch {
-                val id = _category.value?.id ?: return@launch
+                val id = _category.value?.id
+                if (id == null) {
+                    _isDeleted.setValue(false)
+                    return@launch
+                }
+
                 val result: ApiResult<Unit> = categoryRepository.deleteCategory(id)
                 result.onSuccess { updateIsDeleteSuccess() }
                     .onServerError(::handleServerError)
@@ -163,7 +168,7 @@ class CategoryViewModel
         }
 
         private fun updateIsDeleteSuccess() {
-            _isDeleteSuccess.setValue(true)
+            _isDeleted.setValue(true)
         }
 
         private fun handleServerError(message: String) {
